[WIP] [Spike] Gomod major version upgrades
This is a very rough draft of how go major version upgrades could work.
A lot if the implementation was lifted from
https://github.com/marwan-at-work/mod, I've adapted that mostly to fit
in with our slightly different usecase.

The way this works is by after updating the import path in the go.mod
file, we parse all the modules syntax trees and swap out any import
statements.

I opted to do this in a separate step from the existing
"updateDependencyFile" helper, as that returns the new `go.mod` file
rather than update the file in line, and it felt awkward having that
helper change the import paths in the .go files as a side-effect.
Alternatively we could also modify that helper to change the `go.mod`
_and_ the import paths in one go.

There are currently two tests failing which I still need to dive into,
and there are almost certainly edge cases that'll break, so this would
need a bunch of love before being ready to ship, just wanted to share
the proof of concept.

+++ bin/dry-run.rb
@@ -258,9 +258,7 @@ def cached_dependency_files_read
   )
   FileUtils.mkdir_p(cache_dir) unless Dir.exist?(cache_dir)
 
-  if File.exist?(cache_manifest_path)
-    cached_manifest = File.read(cache_manifest_path)
-  end
+  cached_manifest = File.read(cache_manifest_path) if File.exist?(cache_manifest_path)
   cached_dependency_files = JSON.parse(cached_manifest) if cached_manifest
 
   all_files_cached = cached_dependency_files&.all? do |file|
@@ -307,9 +305,7 @@ def cached_dependency_files_read
     end
     # Initialize a git repo so that changed files can be diffed
     if $options[:write]
-      if File.exist?(".gitignore")
-        FileUtils.cp(".gitignore", File.join(cache_dir, ".gitignore"))
-      end
+      FileUtils.cp(".gitignore", File.join(cache_dir, ".gitignore")) if File.exist?(".gitignore")
       Dir.chdir(cache_dir) do
         system("git init . && git add . && git commit --allow-empty -m 'Init'")
       end
@@ -403,19 +399,20 @@ def security_advisories
   end
 end
 
-def peer_dependencies_can_update?(checker, reqs_to_unlock)
-  checker.updated_dependencies(requirements_to_unlock: reqs_to_unlock).
-    reject { |dep| dep.name == checker.dependency.name }.
-    any? do |dep|
-      original_peer_dep = ::Dependabot::Dependency.new(
-        name: dep.name,
-        version: dep.previous_version,
-        requirements: dep.previous_requirements,
-        package_manager: dep.package_manager
-      )
-      update_checker_for(original_peer_dep).
-        can_update?(requirements_to_unlock: :own)
-    end
+def peer_dependencies_can_update?(_checker, _reqs_to_unlock)
+  false
+  # checker.updated_dependencies(requirements_to_unlock: reqs_to_unlock).
+  #   reject { |dep| dep.name == checker.dependency.name }.
+  #   any? do |dep|
+  #     original_peer_dep = ::Dependabot::Dependency.new(
+  #       name: dep.name,
+  #       version: dep.previous_version,
+  #       requirements: dep.previous_requirements,
+  #       package_manager: dep.package_manager
+  #     )
+  #     update_checker_for(original_peer_dep).
+  #       can_update?(requirements_to_unlock: :own)
+  #   end
 end
 
 def file_updater_for(dependencies)
